Fixing a fatal bug and adding new debug.h header
The list_free_all() function in debug.h would stomp over all your
memory and try to free it, because I forgot a *.

The debug.h header provides some macros to print or run debugging
code unless NDEBUG is set (which also disables assert statements).

+++ Makefile
@@ -2,12 +2,18 @@ PREFIX = /usr/local
 LIB_DIR = ${PREFIX}/lib
 INCLUDE_DIR = ${PREFIX}/include
 
-VERSION = 0.1.0
+VERSION = 0.2.0
 
 .PHONY: all install uninstall clean dist
 
 all:
 	make -C src/
+
+dist:
+	install -d dist
+	install -m644 src/*.h dist/
+	install -m644 src/libcassava.so dist/libcassava.so.${VERSION}
+	install -m644 src/libcassava.a dist/
 
 install:
 	install -d ${INCLUDE_DIR}/libcassava
@@ -17,13 +23,14 @@ install:
 	ldconfig ${LIB_DIR}
 
 uninstall:
-	rm -r ${INCLUDE_DIR}/libcassava
-	rm ${LIB_DIR}/libcassava.so.${VERSION}
-	rm ${LIB_DIR}/libcassava.a
+	test -d ${INCLUDE_DIR}/libcassava && rm -r ${INCLUDE_DIR}/libcassava
+	test -f ${LIB_DIR}/libcassava.so.${VERSION} && rm ${LIB_DIR}/libcassava.so.${VERSION}
+	test -f ${LIB_DIR}/libcassava.a && rm ${LIB_DIR}/libcassava.a
 	ldconfig ${LIB_DIR}
 
 clean:
 	make -C src/ clean
-	rm libcassava-${VERSION}.tar.gz
+	test -d dist && rm -rf dist
+	test -w libcassava-${VERSION}.tar.gz && rm libcassava-${VERSION}.tar.gz
 
 # vim: set noet:
